% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_model_performance.R
\name{calculate_model_performance}
\alias{calculate_model_performance}
\title{retrieve vegetation structure data  from NEON}
\usage{
calculate_model_performance(trait = "P_pct", snap_path = "", get_best = F)
}
\value{

}
\description{
retrieve vegetation structure data  from NEON
}
