% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_permutations.R
\name{clean_permutations}
\alias{clean_permutations}
\title{clean the permutation folder from all bag extractions}
\usage{
clean_permutations()
}
\value{

}
\description{
clean the permutation folder from all bag extractions
}
