% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_random_bags.R
\name{get_random_bags}
\alias{get_random_bags}
\title{retrieve vegetation structure data  from NEON}
\usage{
get_random_bags(loops = 1, seed = 1987)
}
\value{

}
\description{
retrieve vegetation structure data  from NEON
}
\examples{
get_random_bags(1000)


}
