% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls_glm_performance.R
\name{pls_glm_performance}
\alias{pls_glm_performance}
\title{retrieve vegetation structure data  from NEON}
\usage{
pls_glm_performance(trait = "N_pct", nbags = 5, normz = F)
}
\value{

}
\description{
retrieve vegetation structure data  from NEON
}
