% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls_glm_predict.R
\name{pls_glm_predict}
\alias{pls_glm_predict}
\title{function inrehited by plsRglm, addind the possibility to calculate probability intervals}
\usage{
pls_glm_predict(
  object,
  newdata,
  comps = object$computed_nt,
  type = c("link", "response", "terms", "scores", "class", "probs"),
  se.fit = FALSE,
  wt = NULL,
  dispersion = NULL,
  methodNA = "adaptative",
  verbose = TRUE,
  ...
)
}
\value{

}
\description{
function inrehited by plsRglm, addind the possibility to calculate probability intervals
}
