% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_model.R
\name{build_model}
\alias{build_model}
\title{retrieve vegetation structure data  from NEON}
\usage{
build_model(loop = 1, nrmlz = T, trait = "P_pct")
}
\value{

}
\description{
retrieve vegetation structure data  from NEON
}
