% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls_glm.R
\name{pls_glm}
\alias{pls_glm}
\title{retrieve vegetation structure data  from NEON}
\usage{
pls_glm(ll = NULL, trait = "N_pct", nrmlz = F)
}
\value{

}
\description{
retrieve vegetation structure data  from NEON
}
